Updating PuzzleSession mechanism. Require to flush all session

+++ applications/accounts/dbroutine.php
@@ -16,12 +16,14 @@ if (file_exists(__ROOTDIR . "/create.admin")) {
 }
 
 #Deleting TFA, and inactive user
-CronJob::register("tfa_remover", function () {
-    Database::execute("DELETE FROM app_users_otp where time < (UNIX_TIMESTAMP() - 600)");
-    Database::execute("DELETE FROM app_users_list where enabled = 0 and registered_time < (UNIX_TIMESTAMP() - 300)");
-}, _CT()->interval(5 * T_MINUTE));
+if (is_cli()) {
+    CronJob::register("tfa_remover", function () {
+        Database::execute("DELETE FROM app_users_otp where time < (UNIX_TIMESTAMP() - 600)");
+        Database::execute("DELETE FROM app_users_list where enabled = 0 and registered_time < (UNIX_TIMESTAMP() - 300)");
+    }, _CT()->interval(5 * T_MINUTE));
+}
 
 #Setting Session
 PuzzleSession::config("retain_on_same_pc", PuzzleUserConfig::enableRememberMe());
 PuzzleSession::config("share_on_subdomain", PuzzleUserConfig::shareSessionToSubdomain());
-PuzzleSession::config("expire",  30 * T_DAY);
+PuzzleSession::config("expire",  7 * T_DAY);
